requirements.txt


Werkzeug==2.0.3
Flask==2.0.3
mysql-connector-python==8.0.29


